# gbs-pos-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Point of Sale REST API for retail, F\u0026B, and fuel station operations",
        "title": "GBS POS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "GBS Support",
            "email": "support@gbs.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/v1/fuel/nozzles": {
            "get": {
                "description": "Get all nozzles",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fuel"
                ],
                "summary": "List nozzles",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new nozzle",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fuel"
                ],
                "summary": "Create nozzle",
                "parameters": [
                    {
                        "description": "Nozzle data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateNozzleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/v1/fuel/nozzles/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a nozzle",
                "tags": [
                    "Fuel"
                ],
                "summary": "Delete nozzle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nozzle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update nozzle",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fuel"
                ],
                "summary": "Update nozzle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nozzle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateNozzleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
        "/v1/fuel/prices/{code}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update fuel price by code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fuel"
                ],
                "summary": "Update fuel price",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fuel code (PERTALITE, PERTAMAX, DEXlite, etc)",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Price update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateFuelPriceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/v1/fuel/pumps": {
            "get": {
                "description": "Get all pumps",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fuel"
                ],
                "summary": "List pumps",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new pump",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fuel"
                ],
                "summary": "Create pump",
                "parameters": [
                    {
                        "description": "Pump data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePumpRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/v1/fuel/pumps/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a pump",
                "tags": [
                    "Fuel"
                ],
                "summary": "Delete pump",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pump ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update pump details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fuel"
                ],
                "summary": "Update pump",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pump ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePumpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/v1/fuel/report": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get fuel sales report for date range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fuel"
                ],
                "summary": "Fuel sales report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/fuel/sales": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Record a fuel sale transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fuel"
                ],
                "summary": "Record fuel sale",
                "parameters": [
                    {
                        "description": "Sale data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.FuelSaleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateNozzleRequest": {
            "type": "object",
            "required": [
                "fuelCode",
                "id",
                "name",
                "pumpId"
            ],
            "properties": {
                "fuelCode": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pumpId": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreatePumpRequest": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.FuelSaleRequest": {
            "type": "object",
            "required": [
                "fuelCode",
                "id",
                "liters",
                "nozzleId",
                "paymentMethod",
                "pricePerLiter",
                "pumpId",
                "totalAmount"
            ],
            "properties": {
                "fuelCode": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "liters": {
                    "type": "number",
                    "minimum": 0
                },
                "nozzleId": {
                    "type": "string"
                },
                "paymentMethod": {
                    "type": "string"
                },
                "posMessageId": {
                    "type": "string"
                },
                "pricePerLiter": {
                    "type": "number",
                    "minimum": 0
                },
                "pumpId": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "totalAmount": {
                    "type": "number",
                    "minimum": 0
                },
                "transactionId": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateFuelPriceRequest": {
            "type": "object",
            "required": [
                "pricePerLiter"
            ],
            "properties": {
                "pricePerLiter": {
                    "type": "number",
                    "minimum": 0
                }
            }
        },
        "dto.UpdateNozzleRequest": {
            "type": "object",
            "properties": {
                "fuelCode": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.UpdatePumpRequest": {
            "type": "object",
            "properties": {
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token. For Keycloak, obtain token from {KEYCLOAK_BASE_URL}/realms/{KEYCLOAK_REALM}/protocol/openid-connect/token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}